provider: second round of reprovider refactor
This revert the revert 4c5c98b94b219d9870c3181f73a99d2f15edae7d and fixes a few deadlocks as well as include a required bump from `go-libp2p-routing-helpers`.
